+++ src/interfaces/packageProviderInterface.ts
@@ -15,4 +15,11 @@ export interface PackageProviderInterface {
      * Returns a package manifest that contains a given file Uri.
      */
     swiftPackageManifestForFile(fileUri: vscode.Uri, cancellation?: vscode.CancellationToken): Promise<SwiftPackageManifest>;
+    
+    /**
+     * Returns a Uri to a package manifest file that contains a given file Uri.
+     * 
+     * May return null, if no package manifest was found containing the file Uri.
+     */
+    swiftPackageManifestPathForFile(fileUri: vscode.Uri, cancellation?: vscode.CancellationToken): Promise<vscode.Uri | null>;
 };