+++ src/test/suite/swiftPackagePathsManager.test.ts
@@ -18,6 +18,7 @@ suite('SwiftPackagePathsManager Test Suite', () => {
             const pkg = makeStandardTestPackage();
 
             const baseUri = vscode.Uri.file(path.join("base", "path"));
+            const manifestUri = vscode.Uri.joinPath(baseUri, "Package.swift");
             const fileUri = vscode.Uri.joinPath(baseUri, "Sources", "file.swift");
 
             fileSystem.createEntriesWithKind(
@@ -25,7 +26,7 @@ suite('SwiftPackagePathsManager Test Suite', () => {
                 [fileUri, VirtualDiskEntryType.file]
             );
 
-            const sut = await SwiftPackagePathsManager.create(baseUri, pkg, fileSystem);
+            const sut = await SwiftPackagePathsManager.create(baseUri, manifestUri, pkg, fileSystem);
 
             assert.strictEqual(await sut.isSourceFile(fileUri), true);
         });
@@ -35,6 +36,7 @@ suite('SwiftPackagePathsManager Test Suite', () => {
             pkg.targets[0].type = TargetType.Executable;
 
             const baseUri = vscode.Uri.file(path.join("base", "path"));
+            const manifestUri = vscode.Uri.joinPath(baseUri, "Package.swift");
             const fileUri = vscode.Uri.joinPath(baseUri, "Sources", "file.swift");
 
             fileSystem.createEntriesWithKind(
@@ -42,7 +44,7 @@ suite('SwiftPackagePathsManager Test Suite', () => {
                 [fileUri, VirtualDiskEntryType.file]
             );
 
-            const sut = await SwiftPackagePathsManager.create(baseUri, pkg, fileSystem);
+            const sut = await SwiftPackagePathsManager.create(baseUri, manifestUri, pkg, fileSystem);
 
             assert.strictEqual(await sut.isSourceFile(fileUri), true);
         });
@@ -52,6 +54,7 @@ suite('SwiftPackagePathsManager Test Suite', () => {
             pkg.targets[0].type = TargetType.Plugin;
 
             const baseUri = vscode.Uri.file(path.join("base", "path"));
+            const manifestUri = vscode.Uri.joinPath(baseUri, "Package.swift");
             const fileUri = vscode.Uri.joinPath(baseUri, "Sources", "file.swift");
 
             fileSystem.createEntriesWithKind(
@@ -59,7 +62,7 @@ suite('SwiftPackagePathsManager Test Suite', () => {
                 [fileUri, VirtualDiskEntryType.file]
             );
 
-            const sut = await SwiftPackagePathsManager.create(baseUri, pkg, fileSystem);
+            const sut = await SwiftPackagePathsManager.create(baseUri, manifestUri, pkg, fileSystem);
 
             assert.strictEqual(await sut.isSourceFile(fileUri), true);
         });
@@ -70,6 +73,7 @@ suite('SwiftPackagePathsManager Test Suite', () => {
             pkg.targets[0].path = "Sources";
 
             const baseUri = vscode.Uri.file(path.join("base", "path"));
+            const manifestUri = vscode.Uri.joinPath(baseUri, "Package.swift");
             const fileUri = vscode.Uri.joinPath(baseUri, "Sources", "file.swift");
 
             fileSystem.createEntriesWithKind(
@@ -77,7 +81,7 @@ suite('SwiftPackagePathsManager Test Suite', () => {
                 [fileUri, VirtualDiskEntryType.file]
             );
 
-            const sut = await SwiftPackagePathsManager.create(baseUri, pkg, fileSystem);
+            const sut = await SwiftPackagePathsManager.create(baseUri, manifestUri, pkg, fileSystem);
 
             assert.strictEqual(await sut.isSourceFile(fileUri), false);
         });
@@ -86,6 +90,7 @@ suite('SwiftPackagePathsManager Test Suite', () => {
             const pkg = makeStandardTestPackage();
 
             const baseUri = vscode.Uri.file(path.join("base", "path"));
+            const manifestUri = vscode.Uri.joinPath(baseUri, "Package.swift");
             const fileUri = vscode.Uri.joinPath(baseUri, "Source", "file.swift");
 
             fileSystem.createEntriesWithKind(
@@ -93,7 +98,7 @@ suite('SwiftPackagePathsManager Test Suite', () => {
                 [fileUri, VirtualDiskEntryType.file]
             );
 
-            const sut = await SwiftPackagePathsManager.create(baseUri, pkg, fileSystem);
+            const sut = await SwiftPackagePathsManager.create(baseUri, manifestUri, pkg, fileSystem);
 
             assert.strictEqual(await sut.isSourceFile(fileUri), true);
         });
@@ -102,6 +107,7 @@ suite('SwiftPackagePathsManager Test Suite', () => {
             const pkg = makeStandardTestPackage();
 
             const baseUri = vscode.Uri.file(path.join("base", "path"));
+            const manifestUri = vscode.Uri.joinPath(baseUri, "Package.swift");
             const fileUri = vscode.Uri.joinPath(baseUri, "src", "file.swift");
 
             fileSystem.createEntriesWithKind(
@@ -109,7 +115,7 @@ suite('SwiftPackagePathsManager Test Suite', () => {
                 [fileUri, VirtualDiskEntryType.file]
             );
 
-            const sut = await SwiftPackagePathsManager.create(baseUri, pkg, fileSystem);
+            const sut = await SwiftPackagePathsManager.create(baseUri, manifestUri, pkg, fileSystem);
 
             assert.strictEqual(await sut.isSourceFile(fileUri), true);
         });
@@ -118,6 +124,7 @@ suite('SwiftPackagePathsManager Test Suite', () => {
             const pkg = makeStandardTestPackage();
 
             const baseUri = vscode.Uri.file(path.join("base", "path"));
+            const manifestUri = vscode.Uri.joinPath(baseUri, "Package.swift");
             const fileUri = vscode.Uri.joinPath(baseUri, "srcs", "file.swift");
 
             fileSystem.createEntriesWithKind(
@@ -125,7 +132,7 @@ suite('SwiftPackagePathsManager Test Suite', () => {
                 [fileUri, VirtualDiskEntryType.file]
             );
 
-            const sut = await SwiftPackagePathsManager.create(baseUri, pkg, fileSystem);
+            const sut = await SwiftPackagePathsManager.create(baseUri, manifestUri, pkg, fileSystem);
 
             assert.strictEqual(await sut.isSourceFile(fileUri), true);
         });
@@ -137,6 +144,7 @@ suite('SwiftPackagePathsManager Test Suite', () => {
             );
 
             const baseUri = vscode.Uri.file(path.join("base", "path"));
+            const manifestUri = vscode.Uri.joinPath(baseUri, "Package.swift");
             const fileUri = vscode.Uri.joinPath(baseUri, "AlternativeSources", "Target", "file.swift");
 
             fileSystem.createEntriesWithKind(
@@ -144,7 +152,7 @@ suite('SwiftPackagePathsManager Test Suite', () => {
                 [fileUri, VirtualDiskEntryType.file]
             );
 
-            const sut = await SwiftPackagePathsManager.create(baseUri, pkg, fileSystem);
+            const sut = await SwiftPackagePathsManager.create(baseUri, manifestUri, pkg, fileSystem);
 
             assert.strictEqual(await sut.isSourceFile(fileUri), true);
         });
@@ -156,6 +164,7 @@ suite('SwiftPackagePathsManager Test Suite', () => {
             );
 
             const baseUri = vscode.Uri.file(path.join("base", "path"));
+            const manifestUri = vscode.Uri.joinPath(baseUri, "Package.swift");
             const fileUri = vscode.Uri.joinPath(baseUri, "AlternativeTests", "TargetTests", "file.swift");
 
             fileSystem.createEntriesWithKind(
@@ -163,7 +172,7 @@ suite('SwiftPackagePathsManager Test Suite', () => {
                 [fileUri, VirtualDiskEntryType.file]
             );
 
-            const sut = await SwiftPackagePathsManager.create(baseUri, pkg, fileSystem);
+            const sut = await SwiftPackagePathsManager.create(baseUri, manifestUri, pkg, fileSystem);
 
             assert.strictEqual(await sut.isSourceFile(fileUri), false);
         });
@@ -175,12 +184,13 @@ suite('SwiftPackagePathsManager Test Suite', () => {
             );
 
             const baseUri = vscode.Uri.file(path.join("base", "path"));
+            const manifestUri = vscode.Uri.joinPath(baseUri, "Package.swift");
 
             fileSystem.createEntriesWithKind(
                 [baseUri, VirtualDiskEntryType.directory],
             );
 
-            const sut = await SwiftPackagePathsManager.create(baseUri, pkg, fileSystem);
+            const sut = await SwiftPackagePathsManager.create(baseUri, manifestUri, pkg, fileSystem);
 
             assert.strictEqual(
                 await sut.isSourceFile(
@@ -202,6 +212,7 @@ suite('SwiftPackagePathsManager Test Suite', () => {
             const pkg = makeStandardTestPackage();
 
             const baseUri = vscode.Uri.file(path.join("base", "path"));
+            const manifestUri = vscode.Uri.joinPath(baseUri, "Package.swift");
             const fileUri = vscode.Uri.joinPath(baseUri, "Tests", "file.swift");
 
             fileSystem.createEntriesWithKind(
@@ -209,7 +220,7 @@ suite('SwiftPackagePathsManager Test Suite', () => {
                 [fileUri, VirtualDiskEntryType.file]
             );
 
-            const sut = await SwiftPackagePathsManager.create(baseUri, pkg, fileSystem);
+            const sut = await SwiftPackagePathsManager.create(baseUri, manifestUri, pkg, fileSystem);
 
             assert.strictEqual(await sut.isTestFile(fileUri), true);
         });
@@ -218,6 +229,7 @@ suite('SwiftPackagePathsManager Test Suite', () => {
             const pkg = makeStandardTestPackage();
 
             const baseUri = vscode.Uri.file(path.join("base", "path"));
+            const manifestUri = vscode.Uri.joinPath(baseUri, "Package.swift");
             const fileUri = vscode.Uri.joinPath(baseUri, "Sources", "file.swift");
 
             fileSystem.createEntriesWithKind(
@@ -225,7 +237,7 @@ suite('SwiftPackagePathsManager Test Suite', () => {
                 [fileUri, VirtualDiskEntryType.file]
             );
 
-            const sut = await SwiftPackagePathsManager.create(baseUri, pkg, fileSystem);
+            const sut = await SwiftPackagePathsManager.create(baseUri, manifestUri, pkg, fileSystem);
 
             assert.strictEqual(await sut.isTestFile(fileUri), true);
         });
@@ -234,6 +246,7 @@ suite('SwiftPackagePathsManager Test Suite', () => {
             const pkg = makeStandardTestPackage();
 
             const baseUri = vscode.Uri.file(path.join("base", "path"));
+            const manifestUri = vscode.Uri.joinPath(baseUri, "Package.swift");
             const fileUri = vscode.Uri.joinPath(baseUri, "Source", "file.swift");
 
             fileSystem.createEntriesWithKind(
@@ -241,7 +254,7 @@ suite('SwiftPackagePathsManager Test Suite', () => {
                 [fileUri, VirtualDiskEntryType.file]
             );
 
-            const sut = await SwiftPackagePathsManager.create(baseUri, pkg, fileSystem);
+            const sut = await SwiftPackagePathsManager.create(baseUri, manifestUri, pkg, fileSystem);
 
             assert.strictEqual(await sut.isTestFile(fileUri), true);
         });
@@ -250,6 +263,7 @@ suite('SwiftPackagePathsManager Test Suite', () => {
             const pkg = makeStandardTestPackage();
 
             const baseUri = vscode.Uri.file(path.join("base", "path"));
+            const manifestUri = vscode.Uri.joinPath(baseUri, "Package.swift");
             const fileUri = vscode.Uri.joinPath(baseUri, "src", "file.swift");
 
             fileSystem.createEntriesWithKind(
@@ -257,7 +271,7 @@ suite('SwiftPackagePathsManager Test Suite', () => {
                 [fileUri, VirtualDiskEntryType.file]
             );
 
-            const sut = await SwiftPackagePathsManager.create(baseUri, pkg, fileSystem);
+            const sut = await SwiftPackagePathsManager.create(baseUri, manifestUri, pkg, fileSystem);
 
             assert.strictEqual(await sut.isTestFile(fileUri), true);
         });
@@ -266,6 +280,7 @@ suite('SwiftPackagePathsManager Test Suite', () => {
             const pkg = makeStandardTestPackage();
 
             const baseUri = vscode.Uri.file(path.join("base", "path"));
+            const manifestUri = vscode.Uri.joinPath(baseUri, "Package.swift");
             const fileUri = vscode.Uri.joinPath(baseUri, "srcs", "file.swift");
 
             fileSystem.createEntriesWithKind(
@@ -273,7 +288,7 @@ suite('SwiftPackagePathsManager Test Suite', () => {
                 [fileUri, VirtualDiskEntryType.file]
             );
 
-            const sut = await SwiftPackagePathsManager.create(baseUri, pkg, fileSystem);
+            const sut = await SwiftPackagePathsManager.create(baseUri, manifestUri, pkg, fileSystem);
 
             assert.strictEqual(await sut.isTestFile(fileUri), true);
         });
@@ -285,6 +300,7 @@ suite('SwiftPackagePathsManager Test Suite', () => {
             );
 
             const baseUri = vscode.Uri.file(path.join("base", "path"));
+            const manifestUri = vscode.Uri.joinPath(baseUri, "Package.swift");
             const fileUri = vscode.Uri.joinPath(baseUri, "AlternativeTests", "TargetTests", "file.swift");
 
             fileSystem.createEntriesWithKind(
@@ -292,7 +308,7 @@ suite('SwiftPackagePathsManager Test Suite', () => {
                 [fileUri, VirtualDiskEntryType.file]
             );
 
-            const sut = await SwiftPackagePathsManager.create(baseUri, pkg, fileSystem);
+            const sut = await SwiftPackagePathsManager.create(baseUri, manifestUri, pkg, fileSystem);
 
             assert.strictEqual(await sut.isTestFile(fileUri), true);
         });
@@ -304,6 +320,7 @@ suite('SwiftPackagePathsManager Test Suite', () => {
             );
 
             const baseUri = vscode.Uri.file(path.join("base", "path"));
+            const manifestUri = vscode.Uri.joinPath(baseUri, "Package.swift");
             const fileUri = vscode.Uri.joinPath(baseUri, "AlternativeSources", "Target", "file.swift");
 
             fileSystem.createEntriesWithKind(
@@ -311,7 +328,7 @@ suite('SwiftPackagePathsManager Test Suite', () => {
                 [fileUri, VirtualDiskEntryType.file]
             );
 
-            const sut = await SwiftPackagePathsManager.create(baseUri, pkg, fileSystem);
+            const sut = await SwiftPackagePathsManager.create(baseUri, manifestUri, pkg, fileSystem);
 
             assert.strictEqual(await sut.isTestFile(fileUri), false);
         });
@@ -323,12 +340,13 @@ suite('SwiftPackagePathsManager Test Suite', () => {
             );
 
             const baseUri = vscode.Uri.file(path.join("base", "path"));
+            const manifestUri = vscode.Uri.joinPath(baseUri, "Package.swift");
 
             fileSystem.createEntriesWithKind(
                 [baseUri, VirtualDiskEntryType.directory],
             );
 
-            const sut = await SwiftPackagePathsManager.create(baseUri, pkg, fileSystem);
+            const sut = await SwiftPackagePathsManager.create(baseUri, manifestUri, pkg, fileSystem);
 
             assert.strictEqual(
                 await sut.isTestFile(